marketdata/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
yfinance>=0.2.40
python-dotenv>=1.0.0
